requirements.txt
fastapi
uvicorn[standard]
streamlit
pandas==2.0.3
numpy==1.24.4
requests
psycopg2-binary
python-dotenv
SQLAlchemy==1.4.52
apache-airflow==2.9.1
python-multipart

